Downcast on GeoTable and re-expose flatgeobuf-wasm (#382)
### Change list

- Start implementing a `Downcast` trait to coerce e.g.
MixedGeometryArrays into a simpler array type if possible. The tricky
part is how to handle chunked geometry arrays. You can't downcast them
_independently_, because you might return a chunked array that's
partially points and partially polygons. For now, I try to check the
data types of each array that it can _downcast to_ and if those are the
same type, I downcast to them.

This can definitely get more sophisticated in the future, but good
enough for now.
- Add `to_array_ref` to convert to `Arc<dyn Array>` from only the trait
object
- Use conditional compilation to build only I/O features in
flatgeobuf-wasm, and exclude data, vector classes, and exclude
algorithms.
- Restore flatgeobuf JS reader 
- Expose `intoTable` to convert a `GeoTable` into a normal Arrow Wasm
Table
- Bump JS version to a beta so I can test release flatgeobuf-wasm