Big table builder revamp! (#391)
I realized with a test FlatGeobuf file that Flatgeobuf also doesn't
mandate including null fields in each record. So we need a builder that
manages checking if any fields were not seen in the current row and
adding null values for those. In this case, it really made sense to
unify the existing flatgeobuf table builder with the schema-less
builders.

This PR adds a generic `GeoTableBuilder<G: GeometryArrayBuilder>` that
can be either schema-less or with a known schema.

We are **2.8x faster** than GDAL's Arrow binding on an example
FlatGeobuf file (without a known geometry type iirc, we should be faster
when the geometry type is stored) and **4.5x faster** than GDAL's Arrow
binding on the same file in GeoJSON!!! 🚀

```
path = '/Users/kyle/Downloads/UScounties.fgb'
arrow_to_pandas_kwargs = {
    'types_mapper': lambda pa_dtype: pd.ArrowDtype(pa_dtype)
}
%timeit geotable = geoarrow.rust.core.read_flatgeobuf(path)
# 8.2 ms ± 276 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

%timeit gdf = pyogrio.read_dataframe(path, use_arrow=True, arrow_to_pandas_kwargs=arrow_to_pandas_kwargs)
# 23.5 ms ± 487 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

```
path = '/Users/kyle/Downloads/UScounties.geojson'
arrow_to_pandas_kwargs = {
    'types_mapper': lambda pa_dtype: pd.ArrowDtype(pa_dtype)
}
%timeit geotable = geoarrow.rust.core.read_geojson(path)
# 175 ms ± 636 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

%timeit gdf = pyogrio.read_dataframe(path, use_arrow=True, arrow_to_pandas_kwargs=arrow_to_pandas_kwargs)
# 795 ms ± 4.72 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### Follow ups:

- Schema handling/resolution among batches

Closes https://github.com/geoarrow/geoarrow-rs/issues/389