Work around upstream compiler bug (#956)
Closes https://github.com/geoarrow/geoarrow-rs/issues/716. Ref
https://github.com/georust/geo/pull/1255#issuecomment-2557320731.

Note that this is the same underlying implementation as upstream geo in
<https://github.com/georust/geo/pull/1255>. However, the trait-based
implementation hits this compiler regression
<https://github.com/rust-lang/rust/issues/128887>,
<https://github.com/rust-lang/rust/issues/131960>, which prevents from
compiling in release mode on a stable Rust version. For some reason, the
**function-based implementation** does not hit this regression, and thus
allows building geoarrow without using latest nightly and a custom
`RUSTFLAGS`.

Note that it's only `GeometryTrait` and `GeometryCollectionTrait` that
hit this compiler bug.
Other traits can use the upstream impls.