Bump to edition 2024 (#1028)
This bumps geoarrow-rs to use edition 2024, see
https://doc.rust-lang.org/edition-guide/rust-2024/index.html and
https://blog.rust-lang.org/2025/02/20/Rust-1.85.0.html. Note that this
causes an **MSRV bump to rust 1.85**.

In general, an MSRV bump to the second most recent rust version wouldn't
be acceptable to many stable Rust libraries, but let's face it,
geoarrow-rs isn't currently in the category of stable rust library. We
can call ourselves bleeding edge for now and take the DX wins of edition
2024. And by the time that geoarrow-rs is more stable (a couple months?)
it'll be less of a hurdle to require Rust 1.85 for all users.

The _second run_ (so compile is fully cached) of running `cargo test
--all-features` **dropped from 76.52s to 14.26s**. And given that we'd
like to add more and more doctests, this on its own is a very nice
improvement to developer experience.