Add cast and downcast methods to Python array and chunked array classes (#1084)
Post https://github.com/geoarrow/geoarrow-rs/pull/1076 and
https://github.com/geoarrow/geoarrow-rs/pull/1080, we now have casting
and downcasting implemented in the core Rust geoarrow-cast crate.

This PR exposes the core `cast` and `downcast` methods to Python.

The Python bindings are still in a higher state of flux so this is just
ensuring that we can pipe the underlying function through.