feat(geoarrow-geo): Initial geoarrow-geo crate using `ArrayAccessor` trait (#1025)
### Highlights

- Starts `geoarrow-geo` crate.
- Pure function might provide simpler implementation than trait methods.
    

[`arrow::compute`](https://docs.rs/arrow/latest/arrow/compute/index.html)
exclusively uses pure functions and not traits, so this might be a good
design decision to copy. I think the main reason why we didn't use pure
functions before was that we wanted to apply the same name onto both
`Array` and `ChunkedArray`. But at this point I think it's better to
leave chunked arrays out of `geoarrow-rs` entirely and leave
chunked/streaming operations to datafusion.
- 

- This defers to the `geo::Area`
[implementation](https://docs.rs/geo/latest/geo/algorithm/area/trait.Area.html#tymethod.unsigned_area).
Note that it uses the upstream implementation [on
`Geometry`](https://docs.rs/geo/latest/geo/algorithm/area/trait.Area.html#impl-Area%3CT%3E-for-Geometry%3CT%3E).
So we don't have to handle the switch on geometry type ourselves because
`geo` will do it for us.
- However, not every geo algorithm is defined on `Geometry`. For
example,
[simplify](https://docs.rs/geo/latest/geo/algorithm/simplify/trait.Simplify.html#implementors)
is defined on lines and polygons but not an arbitrary `Geometry`. This
means that we'd still have to manage downcasting from `Geometry`
ourselves.
- I added a quick example of wrapping simplify. It's still a bit verbose
because we need to figure out what array type to return.
- Shows an example of "specialization". If we know that some geometry
types can have a fast path, then