feat(geodatafusion): Allow loading WKB from BinaryView columns (#1239)
`ST_GeomFromWKB` currently only works with a subset of the standard
binary types. Most users on the latest DataFusion release probably won't
have an issue with this today, but I'm currently using the `main` branch
of DataFusion for several reasons and this defaults to using the view
types (`BinaryView`, `Utf8View`, etc.) instead of `Binary`, `Utf8`, and
so on.

This PR adds support for loading WKB geometries from a `BinaryView`
column, which is how geoparquet geometry fields seem to be interpreted
by default in the upcoming release.