refactor(geoarrow-flatgeobuf): Improved API for passing in or inferring schema (#1272)
### Change list

- Create `FlatGeobufHeaderExt` trait to add new methods to infer
GeoArrow representation
- Allow user to pass in properties schema
- Allows for user to apply projections, change desired data types. Names
must match.
- Rename `FlatGeobufSchemaBuilder` to `FlatGeobufSchemaScanner`
- Allow building WKB directly from FlatGeobuf
- Update datafusion binding

Steps:

1. Open up a FlatGeobuf reader, either sync or async.
2. Use `FlatGeobufHeaderExt` trait to infer an Arrow schema from the
properties + the geometry type
3. If needed, if there's no property information in the header metadata,
open up a new reader on the same file to scan features to infer a
schema. This is the core reason why we need to separate reading the
schema from reading the data, because the user here might need to
construct their own reader to scan the data.
4. Create a `FlatGeobufReaderOptions` with the desired output schema.
Note that you can apply a projection here and change
string/binary/timestamp data types if desired
5. Pass to `FlatGeobufRecordBatchIterator` or
`FlatGeobufRecordBatchStream` and read data.

### Todo

- [x] Write documentation on how to use geoarrow-flatgeobuf
- [ ] Should we actually follow the geometry name on the `schema` passed
in to the options? (added doc note about this)