feat(geoarrow-geojson): GeoJSON writer using arrow-json's new custom encoders (#1040)
Closes https://github.com/geoarrow/geoarrow-rs/issues/982

I wanted to play around with the new functionality in arrow 55 enabled
by https://github.com/apache/arrow-rs/pull/7015.

### Change list

- Creates new `geoarrow-geojson` subcrate, with an additional dependency
on `arrow-json`.
- Adds custom encoders for geometry typed columns.
- Does not currently implement reading from GeoJSON. We can re-enable
that in the future when we come back to a generic table builder API.

Based on upstream benchmarks, this should be _much_ faster than the
existing serde_json-based writing.

### Todo

- [x] Finish implementing all custom encoders for all geometry types.
- [x] Implement encoder for Rect geometry type
- [x] Implement wrapping property columns into a struct column named
`properties`
- [x] Leave `id` column at top level.
- [x] Add tests for all geometry types. Test that the `geojson` crate
can read the types back to ensure that we're writing valid GeoJSON.
    - [ ] Add geometry collection tests
    - [ ] Add null geometry tests
    - [ ] POINT EMPTY